Renamed landing component to landingPage
and fixed build error on import
